# [fingerprint] add useRNCoreAutolinkingFromExpo to support new autolinking