--- a/packages/@expo/fingerprint/build/Config.js
+++ b/packages/@expo/fingerprint/build/Config.js
@@ -41,6 +41,7 @@ async function loadConfigAsync(projectRoot, silent = false) {
         'extraSources',
         'sourceSkips',
         'enableReactImportsPatcher',
+        'useRNCoreAutolinkingFromExpo',
         'debug',
     ];
     const config = {};