--- a/packages/@expo/fingerprint/build/Fingerprint.types.d.ts
+++ b/packages/@expo/fingerprint/build/Fingerprint.types.d.ts
@@ -83,6 +83,11 @@ export interface Options {
      */
     enableReactImportsPatcher?: boolean;
     /**
+     * Use the react-native core autolinking sources from expo-modules-autolinking rather than @react-native-community/cli.
+     * @default true for Expo SDK 52 and higher.
+     */
+    useRNCoreAutolinkingFromExpo?: boolean;
+    /**
      * Whether running the functions should mute all console output. This is useful when fingerprinting is being done as
      * part of a CLI that outputs a fingerprint and outputting anything else pollutes the results.
      */
@@ -96,7 +101,7 @@ type SourceSkipsKeys = keyof typeof SourceSkips;
 /**
  * Supported options from fingerprint.config.js
  */
-export type Config = Pick<Options, 'concurrentIoLimit' | 'hashAlgorithm' | 'ignorePaths' | 'extraSources' | 'enableReactImportsPatcher' | 'debug'> & {
+export type Config = Pick<Options, 'concurrentIoLimit' | 'hashAlgorithm' | 'ignorePaths' | 'extraSources' | 'enableReactImportsPatcher' | 'useRNCoreAutolinkingFromExpo' | 'debug'> & {
     sourceSkips?: SourceSkips | SourceSkipsKeys[];
 };
 export type NormalizedOptions = Omit<Options, 'ignorePaths'> & {