--- a/packages/@expo/fingerprint/build/sourcer/Expo.d.ts
+++ b/packages/@expo/fingerprint/build/sourcer/Expo.d.ts
@@ -1,12 +1,12 @@
 import type { HashSource, NormalizedOptions } from '../Fingerprint.types';
 export declare function getExpoConfigSourcesAsync(projectRoot: string, options: NormalizedOptions): Promise<HashSource[]>;
 export declare function getEasBuildSourcesAsync(projectRoot: string, options: NormalizedOptions): Promise<HashSource[]>;
-export declare function getExpoAutolinkingAndroidSourcesAsync(projectRoot: string, options: NormalizedOptions): Promise<HashSource[]>;
+export declare function getExpoAutolinkingAndroidSourcesAsync(projectRoot: string, options: NormalizedOptions, resolveExpoAutolinkingVersion: string): Promise<HashSource[]>;
 /**
  * Gets the patch sources for the `patch-project`.
  */
 export declare function getExpoCNGPatchSourcesAsync(projectRoot: string, options: NormalizedOptions): Promise<HashSource[]>;
-export declare function getExpoAutolinkingIosSourcesAsync(projectRoot: string, options: NormalizedOptions): Promise<HashSource[]>;
+export declare function getExpoAutolinkingIosSourcesAsync(projectRoot: string, options: NormalizedOptions, expoAutolinkingVersion: string): Promise<HashSource[]>;
 /**
  * Sort the expo-modules-autolinking android config to make it stable from hashing.
  */