--- a/packages/@expo/fingerprint/build/sourcer/ExpoVersions.d.ts
+++ b/packages/@expo/fingerprint/build/sourcer/ExpoVersions.d.ts
@@ -1,0 +1,1 @@
+export declare function resolveExpoAutolinkingVersion(projectRoot: string): string | null;