--- a/packages/@expo/fingerprint/src/sourcer/Bare.ts
+++ b/packages/@expo/fingerprint/src/sourcer/Bare.ts
@@ -72,44 +72,125 @@ export async function getGitIgnoreSourcesAsync(projectRoot: string, options: Nor
   return [];
 }
 
-export async function getRncliAutolinkingSourcesAsync(
+export async function getCoreAutolinkingSourcesFromRncCliAsync(
   projectRoot: string,
-  options: NormalizedOptions
+  options: NormalizedOptions,
+  useRNCoreAutolinkingFromExpo?: boolean
 ): Promise<HashSource[]> {
+  if (useRNCoreAutolinkingFromExpo === true) {
+    return [];
+  }
   try {
-    const results: HashSource[] = [];
     const { stdout } = await spawnAsync('npx', ['react-native', 'config'], { cwd: projectRoot });
     const config = JSON.parse(stdout);
-    const { root } = config;
-    const reasons = ['bareRncliAutolinking'];
-    const autolinkingConfig: Record<string, any> = {};
-    for (const [depName, depData] of Object.entries<any>(config.dependencies)) {
-      try {
-        stripRncliAutolinkingAbsolutePaths(depData, root);
-        const filePath = depData.root;
-        debug(`Adding react-native-cli autolinking dir - ${chalk.dim(filePath)}`);
-        results.push({ type: 'dir', filePath, reasons });
-
-        autolinkingConfig[depName] = depData;
-      } catch (e) {
-        debug(chalk.red(`Error adding react-native-cli autolinking dir - ${depName}.\n${e}`));
-      }
-    }
-
-    results.push({
-      type: 'contents',
-      id: 'rncliAutolinkingConfig',
-      contents: JSON.stringify(autolinkingConfig),
-      reasons,
+    const results: HashSource[] = await parseCoreAutolinkingSourcesAsync({
+      config,
+      contentsId: 'rncoreAutolinkingConfig',
+      reasons: ['rncoreAutolinking'],
     });
     return results;
   } catch (e) {
-    debug(chalk.red(`Error adding react-native-cli autolinking sources.\n${e}`));
+    debug(chalk.red(`Error adding react-native core autolinking sources.\n${e}`));
     return [];
   }
 }
 
-function stripRncliAutolinkingAbsolutePaths(dependency: any, root: string): void {
+export async function getCoreAutolinkingSourcesFromExpoAndroid(
+  projectRoot: string,
+  options: NormalizedOptions,
+  useRNCoreAutolinkingFromExpo?: boolean
+): Promise<HashSource[]> {
+  if (useRNCoreAutolinkingFromExpo === false || !options.platforms.includes('android')) {
+    return [];
+  }
+  try {
+    const { stdout } = await spawnAsync(
+      'npx',
+      ['expo-modules-autolinking', 'react-native-config', '--json', '--platform', 'android'],
+      { cwd: projectRoot }
+    );
+    const config = JSON.parse(stdout);
+    const results: HashSource[] = await parseCoreAutolinkingSourcesAsync({
+      config,
+      contentsId: 'rncoreAutolinkingConfig:android',
+      reasons: ['rncoreAutolinkingAndroid'],
+      platform: 'android',
+    });
+    return results;
+  } catch (e) {
+    debug(chalk.red(`Error adding react-native core autolinking sources for android.\n${e}`));
+    return [];
+  }
+}
+
+export async function getCoreAutolinkingSourcesFromExpoIos(
+  projectRoot: string,
+  options: NormalizedOptions,
+  useRNCoreAutolinkingFromExpo?: boolean
+): Promise<HashSource[]> {
+  if (useRNCoreAutolinkingFromExpo === false || !options.platforms.includes('ios')) {
+    return [];
+  }
+  try {
+    const { stdout } = await spawnAsync(
+      'npx',
+      ['expo-modules-autolinking', 'react-native-config', '--json', '--platform', 'ios'],
+      { cwd: projectRoot }
+    );
+    const config = JSON.parse(stdout);
+    const results: HashSource[] = await parseCoreAutolinkingSourcesAsync({
+      config,
+      contentsId: 'rncoreAutolinkingConfig:ios',
+      reasons: ['rncoreAutolinkingIos'],
+      platform: 'ios',
+    });
+    return results;
+  } catch (e) {
+    debug(chalk.red(`Error adding react-native core autolinking sources for ios.\n${e}`));
+    return [];
+  }
+}
+
+async function parseCoreAutolinkingSourcesAsync({
+  config,
+  reasons,
+  contentsId,
+  platform,
+}: {
+  config: any;
+  reasons: string[];
+  contentsId: string;
+  platform?: string;
+}): Promise<HashSource[]> {
+  const logTag = platform
+    ? `react-native core autolinking dir for ${platform}`
+    : 'react-native core autolinking dir';
+  const results: HashSource[] = [];
+  const { root } = config;
+  const autolinkingConfig: Record<string, any> = {};
+  for (const [depName, depData] of Object.entries<any>(config.dependencies)) {
+    try {
+      stripRncoreAutolinkingAbsolutePaths(depData, root);
+      const filePath = depData.root;
+      debug(`Adding ${logTag} - ${chalk.dim(filePath)}`);
+      results.push({ type: 'dir', filePath, reasons });
+
+      autolinkingConfig[depName] = depData;
+    } catch (e) {
+      debug(chalk.red(`Error adding ${logTag} - ${depName}.\n${e}`));
+    }
+  }
+
+  results.push({
+    type: 'contents',
+    id: contentsId,
+    contents: JSON.stringify(autolinkingConfig),
+    reasons,
+  });
+  return results;
+}
+
+function stripRncoreAutolinkingAbsolutePaths(dependency: any, root: string): void {
   assert(dependency.root);
   const dependencyRoot = dependency.root;
   dependency.root = path.relative(root, dependencyRoot);