… packages/@expo/fingerprint/CHANGELOG.md
… packages/@expo/fingerprint/build/Config.js.map
… packages/@expo/fingerprint/build/Dedup.js.map
… packages/@expo/fingerprint/build/Fingerprint.types.js.map
… packages/@expo/fingerprint/build/sourcer/Bare.js.map
… packages/@expo/fingerprint/build/sourcer/Expo.js.map
… packages/@expo/fingerprint/build/sourcer/ExpoVersions.js.map
… packages/@expo/fingerprint/build/sourcer/Sourcer.js.map
… packages/@expo/fingerprint/e2e/__tests__/__snapshots__/managed-test.ts.snap
… packages/@expo/fingerprint/e2e/__tests__/managed-test.ts
… packages/@expo/fingerprint/src/sourcer/Sourcer.ts
… packages/@expo/fingerprint/src/sourcer/__tests__/Bare-test.ts
… packages/@expo/fingerprint/src/sourcer/__tests__/Expo-test.ts
… packages/@expo/fingerprint/src/sourcer/__tests__/__snapshots__/Bare-test.ts.snap
… packages/@expo/fingerprint/src/sourcer/__tests__/fixtures/RncoreAutoLinkingBadDependencyFromRncCli.json
… packages/@expo/fingerprint/src/sourcer/__tests__/fixtures/RncoreAutoLinkingFromRncCli.json