# Why

since we have expo-modules-autolinking to support the core react-native autolinking sources, we can use it to save some time from fingerprint

# How

add `useRNCoreAutolinkingFromExpo` to support fingerprint

# Benchmark result

bare-expo: `23.048s -> 21.425s`
blank sdk51 project: `881.8ms -> 737.9ms`
react-conf-app: `4.962s -> 2.075s`

# Test Plan

add unit tests and ci passed

# Checklist

- [x] Documentation is up to date to reflect these changes (eg: https://docs.expo.dev and README.md).
- [x] Conforms with the [Documentation Writing Style Guide](https://github.com/expo/expo/blob/main/guides/Expo%20Documentation%20Writing%20Style%20Guide.md)
- [x] This diff will work correctly for `npx expo prebuild` & EAS Build (eg: updated a module plugin).